When making an Item model, initialize description to an empty string.
Previously, some item descriptions were being set to None. Now peewee
thinks that because it has a value, it doesn't set it to our default
value anymore (""). Now instead we add a step where if item.description
is falsey then we initialize it to an empty string. Now SQLite doesn't
complain anymore.